+++ services/users/app/create_user_test.go
@@ -10,8 +10,10 @@ import (
 	"github.com/italoservio/braz_ecommerce/packages/database"
 	"github.com/italoservio/braz_ecommerce/packages/encryption"
 	"github.com/italoservio/braz_ecommerce/services/users/app"
+	"github.com/italoservio/braz_ecommerce/services/users/domain"
 	"github.com/italoservio/braz_ecommerce/services/users/mocks"
 	"github.com/stretchr/testify/assert"
+	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.uber.org/mock/gomock"
 )
 
@@ -44,7 +46,52 @@ func BeforeEach_TestCreateUser(t *testing.T) *TestingDependencies_TestCreateUser
 }
 
 func TestCreateUser_Do(t *testing.T) {
-	t.Run("should return error when failed to call database", func(t *testing.T) {
+	t.Run("should return error when failed to call database CreateOne", func(t *testing.T) {
+		deps := BeforeEach_TestCreateUser(t)
+		defer deps.ctrl.Finish()
+
+		mockExpectedError := errors.New("something goes wrong")
+		mockPassword := "test"
+		mockEmail := "goo@gle.com"
+
+		deps.encryption.
+			EXPECT().
+			Encrypt(gomock.Any(), gomock.Any(), mockPassword).
+			Times(1).
+			Return(&encryption.EncryptedText{EncryptedText: "", Salt: ""}, nil)
+
+		deps.mockUserRepository.
+			EXPECT().
+			GetByEmail(gomock.Any(), database.UsersCollection, mockEmail, gomock.Any()).
+			Times(1).
+			DoAndReturn(func(
+				ctx context.Context,
+				collection string,
+				email string,
+				structure *domain.UserDatabaseNoPassword,
+			) error {
+				*structure = domain.UserDatabaseNoPassword{}
+
+				return nil
+			})
+
+		deps.mockCrudRepository.
+			EXPECT().
+			CreateOne(gomock.Any(), database.UsersCollection, gomock.Any()).
+			Times(1).
+			Return("", mockExpectedError)
+
+		os.Setenv("ENC_SECRET", "2zmXvZa93wneR1w1L63i9cAUzSIzPdd6")
+
+		_, err := deps.createUserImpl.Do(deps.ctx, &app.CreateUserInput{Password: mockPassword, Email: mockEmail})
+		if err == nil {
+			t.Fail()
+		}
+
+		assert.NotNil(t, err, "should return error")
+	})
+
+	t.Run("should return error when calling GetByEmail", func(t *testing.T) {
 		deps := BeforeEach_TestCreateUser(t)
 		defer deps.ctrl.Finish()
 
@@ -57,11 +104,11 @@ func TestCreateUser_Do(t *testing.T) {
 			Times(1).
 			Return(&encryption.EncryptedText{EncryptedText: "", Salt: ""}, nil)
 
-		deps.mockCrudRepository.
+		deps.mockUserRepository.
 			EXPECT().
-			CreateOne(gomock.Any(), database.UsersCollection, gomock.Any()).
+			GetByEmail(gomock.Any(), database.UsersCollection, gomock.Any(), gomock.Any()).
 			Times(1).
-			Return("", mockExpectedError)
+			Return(mockExpectedError)
 
 		os.Setenv("ENC_SECRET", "2zmXvZa93wneR1w1L63i9cAUzSIzPdd6")
 
@@ -71,6 +118,49 @@ func TestCreateUser_Do(t *testing.T) {
 		}
 
 		assert.NotNil(t, err, "should return error")
+	})
+
+	t.Run("should return an error when calling GetByEmail and an email is already registered", func(t *testing.T) {
+		deps := BeforeEach_TestCreateUser(t)
+		defer deps.ctrl.Finish()
+
+		mockPassword := "test"
+		mockEmail := "goo@gle.com"
+
+		deps.encryption.
+			EXPECT().
+			Encrypt(gomock.Any(), gomock.Any(), mockPassword).
+			Times(1).
+			Return(&encryption.EncryptedText{EncryptedText: "", Salt: ""}, nil)
+
+		deps.mockUserRepository.
+			EXPECT().
+			GetByEmail(gomock.Any(), database.UsersCollection, mockEmail, gomock.Any()).
+			Times(1).
+			DoAndReturn(func(
+				ctx context.Context,
+				collection string,
+				email string,
+				structure *domain.UserDatabaseNoPassword,
+			) error {
+				*structure = domain.UserDatabaseNoPassword{
+					DatabaseIdentifier: &database.DatabaseIdentifier{
+						Id: primitive.NewObjectID().Hex(),
+					},
+				}
+
+				return nil
+			})
+
+		os.Setenv("ENC_SECRET", "2zmXvZa93wneR1w1L63i9cAUzSIzPdd6")
+
+		_, err := deps.createUserImpl.Do(deps.ctx, &app.CreateUserInput{Password: mockPassword, Email: mockEmail})
+		if err == nil {
+			t.Fail()
+		}
+
+		assert.NotNil(t, err, "should return error")
+		assert.Equal(t, "EPERMISSION", err.Error(), "should return the expected error code")
 	})
 
 	t.Run("should return error when failed to encrypt password", func(t *testing.T) {
@@ -99,6 +189,22 @@ func TestCreateUser_Do(t *testing.T) {
 		defer deps.ctrl.Finish()
 
 		mockPassword := "test"
+		mockEmail := "goo@gle.com"
+
+		deps.mockUserRepository.
+			EXPECT().
+			GetByEmail(gomock.Any(), database.UsersCollection, mockEmail, gomock.Any()).
+			Times(1).
+			DoAndReturn(func(
+				ctx context.Context,
+				collection string,
+				email string,
+				structure *domain.UserDatabaseNoPassword,
+			) error {
+				*structure = domain.UserDatabaseNoPassword{}
+
+				return nil
+			})
 
 		deps.encryption.
 			EXPECT().
@@ -114,7 +220,7 @@ func TestCreateUser_Do(t *testing.T) {
 
 		os.Setenv("ENC_SECRET", "2zmXvZa93wneR1w1L63i9cAUzSIzPdd6")
 
-		_, err := deps.createUserImpl.Do(deps.ctx, &app.CreateUserInput{Password: mockPassword})
+		_, err := deps.createUserImpl.Do(deps.ctx, &app.CreateUserInput{Password: mockPassword, Email: mockEmail})
 		if err != nil {
 			log.Fatal(err)
 		}