+++ services/users/app/get_user_paginated_test.go
@@ -59,7 +59,7 @@ func TestGetUserPaginated_Do(t *testing.T) {
 			Times(1).
 			Return(mockExpectedError)
 
-		_, err := deps.getUserPaginatedImpl.Do(deps.ctx, input)
+		_, err := deps.getUserPaginatedImpl.Do(deps.ctx, true, input)
 		if err == nil {
 			t.Fail()
 		}
@@ -117,7 +117,66 @@ func TestGetUserPaginated_Do(t *testing.T) {
 				return nil
 			})
 
-		structures, err := deps.getUserPaginatedImpl.Do(deps.ctx, input)
+		structures, err := deps.getUserPaginatedImpl.Do(deps.ctx, true, input)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		assert.Nil(t, err, "should not return an error")
+		assert.Equal(t, 1, len(*structures.Items), "should return one structure")
+	})
+
+	t.Run("should return empty error when executed successfully", func(t *testing.T) {
+		deps := BeforeEach_TestGetUserPaginated(t)
+
+		objId1 := primitive.NewObjectID()
+		objId2 := primitive.NewObjectID()
+
+		input := &app.GetUserPaginatedInput{
+			Page:    1,
+			PerPage: 10,
+			Ids:     []string{objId1.Hex(), objId2.Hex()},
+			Emails:  []string{"foo@bar.net"},
+		}
+
+		filters := map[string]any{
+			"_id":   []primitive.ObjectID{objId1, objId2},
+			"email": []string{"foo@bar.net"},
+		}
+		projection := map[string]int{
+			"password":   0,
+			"cipher_key": 0,
+		}
+		sorting := map[string]int{
+			"created_at": -1,
+		}
+
+		deps.mockCrudRepository.
+			EXPECT().
+			GetPaginated(gomock.Any(), database.UsersCollection, input.Page, input.PerPage, filters, projection, sorting, gomock.Any()).
+			Times(1).
+			DoAndReturn(func(
+				ctx context.Context,
+				collection string,
+				page int,
+				perPage int,
+				filters map[string]any,
+				projection map[string]int,
+				sorting map[string]int,
+				structures any,
+			) error {
+				*structures.(*[]app.GetUserPaginatedOutput) = []app.GetUserPaginatedOutput{
+					{UserDatabaseNoPassword: &domain.UserDatabaseNoPassword{
+						DatabaseIdentifier: &database.DatabaseIdentifier{
+							Id: "123",
+						},
+					}},
+				}
+
+				return nil
+			})
+
+		structures, err := deps.getUserPaginatedImpl.Do(deps.ctx, false, input)
 		if err != nil {
 			log.Fatal(err)
 		}
@@ -135,7 +194,7 @@ func TestGetUserPaginated_Do(t *testing.T) {
 			Ids:     []string{"foo"},
 		}
 
-		_, err := deps.getUserPaginatedImpl.Do(deps.ctx, input)
+		_, err := deps.getUserPaginatedImpl.Do(deps.ctx, true, input)
 		if err == nil {
 			t.Fail()
 		}