+++ services/users/app/update_user_by_id_test.go
@@ -49,6 +49,35 @@ func BeforeEach_TestUpdateUserById(t *testing.T) *TestingDependencies_TestUpdate
 }
 
 func TestUpdateUser_Do(t *testing.T) {
+	t.Run("should return error when failed to call database in GetById", func(t *testing.T) {
+		deps := BeforeEach_TestUpdateUserById(t)
+		defer deps.ctrl.Finish()
+
+		mockEmail := "goo@gle.com"
+
+		mockExpectedError := errors.New("something goes wrong")
+		id := primitive.NewObjectID().Hex()
+
+		deps.mockCrudRepository.
+			EXPECT().
+			GetById(gomock.Any(), database.UsersCollection, id, false, gomock.Any()).
+			Times(1).
+			Return(mockExpectedError)
+
+		deps.mockUserRepository.
+			EXPECT().
+			GetByEmail(gomock.Any(), database.UsersCollection, mockEmail, gomock.Any()).
+			Times(1).
+			Return(mockExpectedError)
+
+		_, err := deps.updateUserByIdImpl.Do(deps.ctx, id, &app.UpdateUserByIdInput{Email: mockEmail})
+		if err == nil {
+			t.Fail()
+		}
+
+		assert.NotNil(t, err, "should return error")
+	})
+
 	t.Run("should return error when failed to call database in GetByEmail", func(t *testing.T) {
 		deps := BeforeEach_TestUpdateUserById(t)
 		defer deps.ctrl.Finish()
@@ -57,6 +86,12 @@ func TestUpdateUser_Do(t *testing.T) {
 
 		mockExpectedError := errors.New("something goes wrong")
 		id := primitive.NewObjectID().Hex()
+
+		deps.mockCrudRepository.
+			EXPECT().
+			GetById(gomock.Any(), database.UsersCollection, id, false, gomock.Any()).
+			Times(1).
+			Return(nil)
 
 		deps.mockUserRepository.
 			EXPECT().
@@ -79,6 +114,12 @@ func TestUpdateUser_Do(t *testing.T) {
 		mockEmail := "goo@gle.com"
 
 		id := primitive.NewObjectID().Hex()
+
+		deps.mockCrudRepository.
+			EXPECT().
+			GetById(gomock.Any(), database.UsersCollection, id, false, gomock.Any()).
+			Times(1).
+			Return(nil)
 
 		deps.mockUserRepository.
 			EXPECT().
@@ -115,6 +156,12 @@ func TestUpdateUser_Do(t *testing.T) {
 		mockEmail := "goo@gle.com"
 
 		id := primitive.NewObjectID().Hex()
+
+		deps.mockCrudRepository.
+			EXPECT().
+			GetById(gomock.Any(), database.UsersCollection, id, false, gomock.Any()).
+			Times(1).
+			Return(nil)
 
 		deps.mockUserRepository.
 			EXPECT().
@@ -160,6 +207,12 @@ func TestUpdateUser_Do(t *testing.T) {
 
 		id := primitive.NewObjectID().Hex()
 
+		deps.mockCrudRepository.
+			EXPECT().
+			GetById(gomock.Any(), database.UsersCollection, id, false, gomock.Any()).
+			Times(1).
+			Return(nil)
+
 		deps.mockUserRepository.
 			EXPECT().
 			GetByEmail(gomock.Any(), database.UsersCollection, mockEmail, gomock.Any()).
@@ -203,6 +256,12 @@ func TestUpdateUser_Do(t *testing.T) {
 		mockEmail := "goo@gle.com"
 
 		id := primitive.NewObjectID().Hex()
+
+		deps.mockCrudRepository.
+			EXPECT().
+			GetById(gomock.Any(), database.UsersCollection, id, false, gomock.Any()).
+			Times(1).
+			Return(nil)
 
 		deps.mockUserRepository.
 			EXPECT().